Flatten and normalize DNA module structure
