+++ core_types/src/dna/entry_types.rs
